Add template override option to frontmatter
