+++ packages/site/plugins/source-markdown/node.api.js
@@ -1,10 +1,5 @@
 import { createSharedData } from 'react-static/node';
-import frontmatter from 'remark-frontmatter';
-import squeeze from 'remark-squeeze-paragraphs';
-import toString from 'mdast-util-to-string';
-import { parse as yaml } from 'yaml';
 import * as path from 'path';
-import remark from 'remark';
 
 import { getPages } from './markdown';
 
@@ -15,8 +10,8 @@ const staticPluginSourceMarkdown = (opts = {}) => ({
 
     // Get page data for each discovered markdown file
     const pages = await getPages(
-      getMarkdownProcessor(opts.remarkPlugins),
       location,
+      opts.remarkPlugins,
       opts.pathPrefix
     );
 
@@ -56,7 +51,7 @@ const staticPluginSourceMarkdown = (opts = {}) => ({
   webpack(webpackConfig, { config, defaultLoaders }) {
     // Resolve target location and template from ROOT folder
     const location = path.resolve(config.paths.ROOT, opts.location);
-    const template = path.resolve(config.paths.ROOT, opts.template);
+    const defaultTemplate = path.resolve(config.paths.ROOT, opts.template);
 
     // Create a rule that only applies to the discovered markdown files
     webpackConfig.module.rules[0].oneOf.unshift({
@@ -70,8 +65,8 @@ const staticPluginSourceMarkdown = (opts = {}) => ({
         {
           loader: require.resolve('./loader'),
           options: {
-            processor: getMarkdownProcessor(opts.remarkPlugins),
-            template,
+            remarkPlugins: opts.remarkPlugins,
+            defaultTemplate,
             location,
           },
         },
@@ -81,24 +76,5 @@ const staticPluginSourceMarkdown = (opts = {}) => ({
     return webpackConfig;
   },
 });
-
-const getMarkdownProcessor = (plugins = []) => {
-  // By default the remark parsers gets the frontmatter data and removes
-  // extra-long paragraphs
-  const processor = remark()
-    .use(frontmatter, ['yaml'])
-    .use(squeeze);
-
-  // All plugins in opts.remarkPlugins will be added to remark
-  plugins.forEach(plugin => {
-    if (Array.isArray(plugin) && plugin.length > 1) {
-      fn.use(plugin[0], plugin[1]);
-    } else {
-      fn.use(plugin);
-    }
-  });
-
-  return processor;
-};
 
 export default staticPluginSourceMarkdown;
